add error message
if tasks in the folder, return error massage and don't delete.

+++ app/Http/Controllers/FolderController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers;
 use App\Models\Folder;
 use Illuminate\Http\Request;
 use App\Http\Requests\CreateFolder;
+use App\Models\Task;
 // ★ Authクラスをインポートする
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
 
 class FolderController extends Controller
 {
@@ -33,9 +35,19 @@ class FolderController extends Controller
         ]);
     }
 
-    public function destroy(Folder $folder){
-
+    public function delete(Folder $folder){
         $deleteid = $folder->id;
+
+        //タスクの存在チェック
+        $taskCount = Task::where('folder_id', $deleteid)->count();
+        if ($taskCount !== 0){
+            $error = "フォルダの中にタスクが存在します";
+            return redirect()->route('tasks.index', [
+                'folder' => $folder->id,
+            ])
+            ->withErrors($error);
+        };
+
         Folder::destroy($deleteid);
         return redirect()->route('home');
     }
